+++ src/matterbridgePlatform.ts
@@ -27,6 +27,8 @@ import path from 'node:path';
 
 // Matter
 import { EndpointNumber } from '@matter/main';
+import { Descriptor } from '@matter/main/clusters/descriptor';
+import { BridgedDeviceBasicInformation } from '@matter/main/clusters/bridged-device-basic-information';
 // Node AnsiLogger module
 import { AnsiLogger, CYAN, db, er, LogLevel, nf, wr } from 'node-ansi-logger';
 // Node Storage module
@@ -36,6 +38,7 @@ import { NodeStorage, NodeStorageManager } from 'node-persist-manager';
 import { Matterbridge } from './matterbridge.js';
 import { MatterbridgeEndpoint } from './matterbridgeEndpoint.js';
 import { checkNotLatinCharacters } from './matterbridgeEndpointHelpers.js';
+import { bridgedNode } from './matterbridgeDeviceTypes.js';
 import { isValidArray, isValidObject, isValidString } from './utils/export.js';
 
 // Platform types
@@ -278,34 +281,83 @@ export class MatterbridgePlatform {
   }
 
   /**
-   * Registers a device with the Matterbridge platform.
+   * Registers a device with the Matterbridge platform and performs validation checks.
+   *
+   * This method also checks if the implementation called createDefaultBasicInformationClusterServer() instead of createDefaultBridgedDeviceBasicInformationClusterServer().
+   *
+   * This is correct with Accessory platforms so we check if we are running in bridge mode and add the bridgedNode and the BridgedDeviceBasicInformation cluster.
+   *
+   * If we are in bridge mode, we add the bridgedNode and the BridgedDeviceBasicInformation cluster.
+   *
+   * If we are in childbridge mode and the plugin is a 'DynamicPlatform', we add the bridgedNode and the BridgedDeviceBasicInformation cluster.
+   *
+   * if we are in childbridge mode and the plugin is a 'AccessoryPlatform', the device is not bridged.
+   *
+   * If the device.mode = 'server', the device is not bridged.
+   *
+   * If the device.mode = 'matter', the device is not bridged.
    *
    * @param {MatterbridgeEndpoint} device - The device to register.
    */
   async registerDevice(device: MatterbridgeEndpoint) {
     device.plugin = this.name;
-    if (device.deviceName && this._registeredEndpointsByName.has(device.deviceName)) {
+    if (!device.uniqueId) {
+      this.log.error(
+        `Device with name ${CYAN}${device.deviceName}${er} has no uniqueId. Did you forget to call createDefaultBasicInformationClusterServer() or createDefaultBridgedDeviceBasicInformationClusterServer()? The device will not be added.`,
+      );
+      return;
+    }
+    if (!device.deviceName) {
+      this.log.error(`Device with uniqueId ${CYAN}${device.uniqueId}${er} has no deviceName. The device will not be added.`);
+      return;
+    }
+    if (!device.serialNumber) {
+      this.log.error(`Device with uniqueId ${CYAN}${device.uniqueId}${er} has no serialNumber. The device will not be added.`);
+      return;
+    }
+    if (this._registeredEndpointsByName.has(device.deviceName)) {
       this.log.error(`Device with name ${CYAN}${device.deviceName}${er} is already registered. The device will not be added. Please change the device name.`);
       return;
     }
-    if (device.deviceName && checkNotLatinCharacters(device.deviceName)) {
+    if (checkNotLatinCharacters(device.deviceName)) {
       this.log.debug(`Device with name ${CYAN}${device.deviceName}${db} has non latin characters.`);
     }
 
-    /* TODO: Do we need to handle the childbridged mode with DynamicPlatform here if the implementation forgets to call createDefaultBridgedDeviceBasicInformationClusterServer?
-    if (this.matterbridge.bridgeMode === 'childbridge' && this.type === 'DynamicPlatform' && device.mode === undefined) {
-      const options = device.getClusterServerOptions(Descriptor.Cluster.id);
-      if (options) {
-        const deviceTypeList = options.deviceTypeList as { deviceType: number; revision: number }[];
-        deviceTypeList.push({ deviceType: bridgedNode.code, revision: bridgedNode.revision });
+    // Validate bridgedNode and BridgedDeviceBasicInformation cluster
+    if (device.mode === undefined && (this.matterbridge.bridgeMode === 'bridge' || (this.matterbridge.bridgeMode === 'childbridge' && this.type === 'DynamicPlatform'))) {
+      // If the device is a bridged device, we add the bridgedNode to the deviceTypes map and to the Descriptor Cluster options
+      if (!device.deviceTypes.has(bridgedNode.code)) {
+        this.log.debug(`Device with name ${CYAN}${device.deviceName}${db} has no bridgedNode device type. Adding it...`);
+        device.deviceTypes.set(bridgedNode.code, bridgedNode);
+        const options = device.getClusterServerOptions(Descriptor.Cluster.id);
+        if (options) {
+          const deviceTypeList = options.deviceTypeList as { deviceType: number; revision: number }[];
+          if (!deviceTypeList.find((dt) => dt.deviceType === bridgedNode.code)) {
+            deviceTypeList.push({ deviceType: bridgedNode.code, revision: bridgedNode.revision });
+          }
+        }
       }
-      device.createDefaultBridgedDeviceBasicInformationClusterServer(deviceName, serialNumber, vendorId, vendorName, productName, softwareVersion, softwareVersionString, hardwareVersion, hardwareVersionString);
+
+      // If the device is a bridged device, we add the BridgedDeviceBasicInformation cluster
+      if (!device.hasClusterServer(BridgedDeviceBasicInformation.Cluster.id)) {
+        this.log.debug(`Device with name ${CYAN}${device.deviceName}${db} has no BridgedDeviceBasicInformation cluster. Adding it...`);
+        device.createDefaultBridgedDeviceBasicInformationClusterServer(
+          device.deviceName,
+          device.serialNumber,
+          device.vendorId,
+          device.vendorName,
+          device.productName,
+          device.softwareVersion,
+          device.softwareVersionString,
+          device.hardwareVersion,
+          device.hardwareVersionString,
+        );
+      }
     }
-    */
 
     await this.matterbridge.addBridgedEndpoint(this.name, device);
-    if (device.uniqueId) this._registeredEndpoints.set(device.uniqueId, device);
-    if (device.deviceName) this._registeredEndpointsByName.set(device.deviceName, device);
+    this._registeredEndpoints.set(device.uniqueId, device);
+    this._registeredEndpointsByName.set(device.deviceName, device);
   }
 
   /**